% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disableSteps.R
\name{disableSteps}
\alias{disableSteps}
\title{disableSteps}
\usage{
disableSteps()
}
\description{
Server logic to disable all step tabs except the 1st tab 
(Select data) at the beginning of the app execution.
}
\author{
Jedid Ahn
}
