% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDictDF.R
\name{getDictDF}
\alias{getDictDF}
\title{getDictDF}
\usage{
getDictDF(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
Dictionary data frame if it exists, otherwise NULL.
}
\description{
Getter function to retrieve dictionary data frame.
}
\author{
Jedid Ahn
}
