% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumComponents.R
\name{getNumComponents}
\alias{getNumComponents}
\title{getNumComponents}
\usage{
getNumComponents(input, rvStep6Results, rvVarDataTypes)
}
\arguments{
\item{input}{Shiny input}

\item{rvStep6Results}{Reactive value list containing transDataDF, asIsDF,
DF_EXCLUDED_COLS}

\item{rvVarDataTypes}{Reactive value list containing catVars, ordVars,
numVars, asIsVars, excludeVars}
}
\value{
List containing max and default number of components to impute on.
}
\description{
Getter function to retrieve the number of components to set
the slider value in step #7.
}
\author{
Jedid Ahn
}
