% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOriginalDataDF.R
\name{getOriginalDataDF}
\alias{getOriginalDataDF}
\title{getOriginalDataDF}
\usage{
getOriginalDataDF(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
Data data frame if it exists, otherwise NULL.
}
\description{
Getter function to retrieve the original data data frame.
}
\author{
Jedid Ahn
}
