% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardCols.R
\name{getStandardCols}
\alias{getStandardCols}
\title{getStandardCols}
\usage{
getStandardCols(dictDF, includeSubject = FALSE)
}
\arguments{
\item{dictDF}{Dictionary data frame.}

\item{includeSubject}{Optional logical parameter, which indicates whether 
SUBJECT column should be included or not (as some data frames will contain
the subject ID in the row names).}
}
\value{
Column numbers corresponding to the ONDRI standard columns. If
dictDF is NULL, return numeric(0).
}
\description{
Getter function to retrieve the following ONDRI standard 
columns: SUBJECT, VISIT, _SITE, _DATE.
}
\author{
Jedid Ahn
}
