% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStep5Results.R
\name{getStep5Results}
\alias{getStep5Results}
\title{getStep5Results}
\usage{
getStep5Results(input, rvOriginalDataDF, rvDictDF, ONDRIMissingCodes)
}
\arguments{
\item{input}{Shiny input}

\item{rvOriginalDataDF}{Reactive value containing the original data
data frame}

\item{rvDictDF}{Reactive value containing the dictionary data frame}

\item{ONDRIMissingCodes}{ONDRI missing codes}
}
\value{
List containing data frame of dropped columns (DF_DROPPED_COLS),
data frame of dropped rows (DF_DROPPED_ROWS), vector of variable names
dropped (varNames), vector of participant names dropped (parNames), and
updated data data frame (dataDF).
}
\description{
Getter function to retrieve the updated data data frame after
step #5 is complete along with the variables and participants that were
dropped.
}
\author{
Jedid Ahn
}
