% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStep6Results.R
\name{getStep6Results}
\alias{getStep6Results}
\title{getStep6Results}
\usage{
getStep6Results(input, rvStep5Results, rvVarDataTypes, rvDictDF)
}
\arguments{
\item{input}{Shiny input}

\item{rvStep5Results}{Reactive value list containing DF_DROPPED_COLS, 
DF_DROPPED_ROWS, varNames, parNames, dataDF}

\item{rvVarDataTypes}{Reactive value list containing catVars, ordVars,
numVars, asIsVars, excludeVars}

\item{rvDictDF}{Reactive value containing the dictionary data frame}
}
\value{
List containing transformed data frame (transDataDF), data frame
of columns to leave as is (asIsDF), and data frame of columns to exclude
(DF_EXCLUDED_COLS).
}
\description{
Getter function to retrieve the updated data data frame after
performing transformations according to the variable types chosen.
}
\author{
Jedid Ahn
}
