% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStep7Results.R
\name{getStep7Results}
\alias{getStep7Results}
\title{getStep7Results}
\usage{
getStep7Results(input, rvStep6Results, rvVarDataTypes)
}
\arguments{
\item{input}{Shiny input}

\item{rvStep6Results}{Reactive value list containing transDataDF, asIsDF,
DF_EXCLUDED_COLS}

\item{rvVarDataTypes}{Reactive value list containing catVars, ordVars,
numVars, asIsVars, excludeVars}
}
\value{
Prepared data to export.
}
\description{
Getter function to retrieve the updated data data frame after
imputation.
}
\author{
Jedid Ahn
}
