% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarDataTypes.R
\name{getVarDataTypes}
\alias{getVarDataTypes}
\title{getVarDataTypes}
\usage{
getVarDataTypes(input, rvStep5Results, rvDictDF)
}
\arguments{
\item{input}{Shiny input}

\item{rvStep5Results}{Reactive value list containing DF_DROPPED_COLS, 
DF_DROPPED_ROWS, varNames, parNames, dataDF}

\item{rvDictDF}{Reactive value containing the dictionary data frame}
}
\value{
List of 5 vectors, where each vector contains variable 
names for a specific data type.
}
\description{
Getter function to retrieve the variable data types after
selection by user in step #6 of the pipeline. The 5 data types are as 
follows: CATEGORICAL, ORDINAL, NUMERIC, AS IS, EXCLUDE
}
\author{
Jedid Ahn
}
