% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1Server.R
\name{step1Server}
\alias{step1Server}
\title{step1Server}
\usage{
step1Server(input, output, session, rvDataDF)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{rvDataDF}{Data file}
}
\description{
Server logic for the step #1 (Select data) tab by reading in
DATA file from the local computer.
}
\author{
Jedid Ahn
}
