% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2Server.R
\name{step2Server}
\alias{step2Server}
\title{step2Server}
\usage{
step2Server(input, output, session, rvDataDF, rvDictDF)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{rvDataDF}{Data file}

\item{rvDictDF}{Dictionary file}
}
\description{
Server logic for the step #2 (Select dictionary) tab by reading
in DICT file from the local computer.
}
\author{
Jedid Ahn
}
