% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2UI.R
\name{step2UI}
\alias{step2UI}
\title{step2UI}
\usage{
step2UI()
}
\description{
User interface for the step #2 (Select dictionary) tab to load
in a data dictionary frame from the local computer.
}
\author{
Jedid Ahn
}
