% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3UI.R
\name{step3UI}
\alias{step3UI}
\title{step3UI}
\usage{
step3UI()
}
\description{
User interface for the step #3 (Visualize missingness by type)
tab to visualize missingness by type through colour-coded heat maps.
}
\author{
Jedid Ahn
}
