% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4Server.R
\name{step4Server}
\alias{step4Server}
\title{step4Server}
\usage{
step4Server(input, output, session, rvDataDF, ONDRIMissingCodes)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{rvDataDF}{Data file}

\item{ONDRIMissingCodes}{ONDRI missing codes}
}
\description{
Server logic for the step #4 (Assess summary of 
missingness) tab.
}
\author{
Jedid Ahn
}
