% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4UI.R
\name{step4UI}
\alias{step4UI}
\title{step4UI}
\usage{
step4UI()
}
\description{
User interface for the step #4 (Assess summary of missingness)
tab to assess summary of missingness through contingency tables of rows
and columns.
}
\author{
Jedid Ahn
}
