% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5UI.R
\name{step5UI}
\alias{step5UI}
\title{step5UI}
\usage{
step5UI()
}
\description{
User interface for the step #5 (Define missingness thresholds) 
tab to select rows and columns to drop from our data frame.
}
\author{
Jedid Ahn
}
