% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6Server.R
\name{step6Server}
\alias{step6Server}
\title{step6Server}
\usage{
step6Server(input, output, session, rvStep5Results, rvStep6Results, rvDictDF)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{rvStep5Results}{Reactive value list containing DF_DROPPED_COLS, 
DF_DROPPED_ROWS, varNames, parNames, dataDF}

\item{rvStep6Results}{Reactive value list containing transDataDF, asIsDF,
DF_EXCLUDED_COLS}

\item{rvDictDF}{Dictionary file}
}
\description{
Server logic for the step #6 (Define variable types and select
variables to drop) tab.
}
\author{
Jedid Ahn
}
