% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6UI.R
\name{step6UI}
\alias{step6UI}
\title{step6UI}
\usage{
step6UI()
}
\description{
User interface for the step #6 (Define variable types and
select variables to drop) tab to allow the user to customize variable 
types in their data.  Variable types include the following: 
TEXT, DATE, CATEGORICAL, ORDINAL, NUMERIC, MIXED.
}
\author{
Jedid Ahn
}
