% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7Server.R
\name{step7Server}
\alias{step7Server}
\title{step7Server}
\usage{
step7Server(
  input,
  output,
  session,
  rvStep5Results,
  rvStep6Results,
  rvStep7Results,
  rvVarDataTypes
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{rvStep5Results}{Reactive value list containing DF_DROPPED_COLS, 
DF_DROPPED_ROWS, varNames, parNames, dataDF}

\item{rvStep6Results}{Reactive value list containing transDataDF, asIsDF,
DF_EXCLUDED_COLS}

\item{rvStep7Results}{Reactive value containing DF_PREPPED_DATA}

\item{rvVarDataTypes}{Reactive value list containing catVars, ordVars,
numVars, asIsVars, excludeVars}
}
\description{
Server logic for the step #7 (Perform imputation) tab.
}
\author{
Jedid Ahn
}
