% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7UI.R
\name{step7UI}
\alias{step7UI}
\title{step7UI}
\usage{
step7UI()
}
\description{
User interface for the step #7 (Perform imputation) tab to 
perform imputation on continuous and non-continuous data.
}
\author{
Jedid Ahn
}
